requirements.txt
requests
feedparser
apscheduler
python-dotenv
google-genai
